Development
- Added `onCollectionCreate` event to `Model` class.
- Renamed `Model/ModelCollection` to `Model/Collections/ModelCollection`.
- Moved Collection-related methods in `ModelNode` to `ModelCollection` class.
- Minor enhancements to `DbTranslateProvider` class.
